adding schema and joi
