+++ models/index.js
@@ -1,0 +1,6 @@
+const Contact = require("./contact");
+
+
+module.export = {
+    Contact
+}
